Add dv displayer

+++ RTProject/src/main/java/helper/Graph.java
@@ -33,7 +33,7 @@ public class Graph {
         
         for (ElkNode n : graph.getChildren())
             g.nodes.add(new Node(n.getX(), n.getY(), n.getWidth(), n.getHeight(), n.getIdentifier(), 
-            		Double.toString(Help.getProp(n).xOffset), new ArrayList<Edge>(), new ArrayList<Edge>(), n, Help.getProp(n).thread));
+            		Integer.toString(Help.getProp(n).xOffset), new ArrayList<Edge>(), new ArrayList<Edge>(), n, Help.getProp(n).thread));
         
         for (ElkEdge e : graph.getContainedEdges())
             g.edges.add(new Edge(
